refactor(attendance): 출결 관련 클래스 리팩터링
- `ClassSchedule` 열거형을 `AttendanceStatus` 클래스에서 분리
